+++ mastermlx/base/__init__.py
@@ -1,9 +1,10 @@
 from .estimator import BaseEstimator
 from .transformer import BaseTransformer
 from .layer import BaseLayer
+from .module import Module, Parameter
 
 Est = BaseEstimator
 Trans = BaseTransformer
 Layer = BaseLayer
 
-__all__ = ["BaseEstimator", "BaseTransformer", "BaseLayer", "Est", "Trans", "Layer"]
+__all__ = ["BaseEstimator", "BaseTransformer", "BaseLayer", "Module", "Parameter", "Est", "Trans", "Layer"]